// src/types/FTTHBlock.ts
export type FTTHBlock = {
    ID: number;
    Name: string;
    Block_ID: number;
    Adres1395: string;
    Hoze1395: number;
    BLK_No1395: number;
    Value: number;
    "95HH": number;
    Area: number;
    length: number;
    Lat: number;
    Long: number;
    Chain_ID: number;
    Order: number;
    City: string;
  };
  